Adds Chord support for the cache backend.
The implementation is efficient and should make chords for the cache
backend just as good, or better as when using the Redis backend.

(dummy cache support has also been added, but being in-memory
this is rather limited and to be used for unittests only)

Also adds unittests and fixes the memcache backend's test reset code:

If one changes the common cache backend now shared in 'class
test_CacheBackend' to memcached instead of the dummy backend (for
example, to actually test this backend), then this code currently fails
as the _imp variable is not reset in this one case. Fix this small
omission and make some of the other code clearer (explicit tuple) and
more consistent (order of the mock context decorators).

Closes #533

+++ celery/backends/redis.py
@@ -74,19 +74,21 @@ class RedisBackend(KeyValueStoreBackend):
 
     def on_chord_apply(self, setid, body, result=None, **kwargs):
         self.app.TaskSetResult(setid, result).save()
-        pass
 
-    def on_chord_part_return(self, task, propagate=False,
-            keyprefix="chord-unlock-%s"):
+    def on_chord_part_return(self, task, propagate=False):
         from ..task.sets import subtask
         from ..result import TaskSetResult
         setid = task.request.taskset
-        key = keyprefix % setid
+        if not setid:
+            return
+        key = self.get_key_for_chord(setid)
         deps = TaskSetResult.restore(setid, backend=task.backend)
         if self.client.incr(key) >= deps.total:
             subtask(task.request.chord).delay(deps.join(propagate=propagate))
             deps.delete()
-        self.client.expire(key, 86400)
+            self.client.delete(key)
+        else:
+            self.client.expire(key, 86400)
 
     @cached_property
     def client(self):
